+++ src/main/java/com/rudkids/rudkids/domain/item/domain/itemOption/ItemOptionStore.java
@@ -1,4 +1,4 @@
-package com.rudkids.rudkids.domain.item.domain.itemOptionGroup.itemOption;
+package com.rudkids.rudkids.domain.item.domain.itemOption;
 
 public interface ItemOptionStore {
     ItemOption store(ItemOption itemOption);
