+++ src/main/java/com/rudkids/rudkids/infrastructure/item/itemOptionGroup/itemOption/ItemOptionRepository.java
@@ -1,6 +1,6 @@
 package com.rudkids.rudkids.infrastructure.item.itemOptionGroup.itemOption;
 
-import com.rudkids.rudkids.domain.item.domain.itemOptionGroup.itemOption.ItemOption;
+import com.rudkids.rudkids.domain.item.domain.itemOption.ItemOption;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
